avances en enviar aviso de incidencia 3
Falta almacenar el archivo adjunto y mostrar algún mensaje de "ha
ocurrido satisfactoriamente..."

+++ coplime/coplime-ejb/src/java/entities/Administrador.java
@@ -34,23 +34,23 @@ public class Administrador implements Serializable {
         this.usuario = usuario;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
     
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
-    public Long getCod() {
+    public Integer getCod() {
         return id;
     }
 
-    public void setCod(Long cod) {
+    public void setCod(Integer cod) {
         this.id = cod;
     }
