+++ coplime/coplime-ejb/src/java/entities/OperarioMantencion.java
@@ -45,24 +45,24 @@ public class OperarioMantencion implements Serializable {
     
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
-    public Long getCod() {
+    public Integer getCod() {
         return id;
     }
 
-    public void setCod(Long cod) {
+    public void setCod(Integer cod) {
         this.id = cod;
     }
     
     @OneToOne
     private Usuario usuario;
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
